RNP-80: Rename flow files to use<type>Flow
